+++ app/src/main/java/com/github/versus/user/VersusUser.java
@@ -27,7 +27,7 @@ public final class VersusUser implements User, Serializable {
     private final String userName;
     private final String mail;
     private final String phone;
-    private final int rating;
+    private final long rating;
     private final String city;
     private final int zip;
     private final List<Sport> preferredSports;
@@ -45,7 +45,7 @@ public final class VersusUser implements User, Serializable {
         firstName = "";
         lastName = "";
     }
-    public VersusUser(String uid, String firstName,String lastName,String userName, String mail, String phone , int rating , String city, int zipCode,List<Sport> preferredSports ){
+    public VersusUser(String uid, String firstName,String lastName,String userName, String mail, String phone , long rating , String city, int zipCode,List<Sport> preferredSports ){
         this.uid = uid;
         this.firstName =firstName;
         this.lastName = lastName;
@@ -101,7 +101,7 @@ public final class VersusUser implements User, Serializable {
     }
 
     @Override
-    public int getRating() {
+    public long getRating() {
         return rating;
     }
 
@@ -150,7 +150,7 @@ public final class VersusUser implements User, Serializable {
         private String userName;
         private String mail;
         private String phone;
-        private int rating;
+        private long rating;
         private String city;
         private int zipCode;
         private List<Sport> preferredSports = new ArrayList<>();
@@ -224,7 +224,7 @@ public final class VersusUser implements User, Serializable {
          * @return
          */
         @Override
-        public VersusBuilder setRating(int rating){
+        public VersusBuilder setRating(long rating){
             this.rating = rating;
             return this;
         }
@@ -264,7 +264,7 @@ public final class VersusUser implements User, Serializable {
 
         /**
          * ???
-         * @param userName
+         * @param
          * @return
          */
         @Override
